requirements.txt
PySide6>=6.6
numpy>=1.24
opencv-python>=4.8
pyserial>=3.5
tifffile>=2023.7.10
PyYAML>=6.0
pytest>=8.0
